Fix errors converting empty values in forms.
The removed default converters for number types
did not have special logic for handling the empty
string leading to NumberFormatExceptions.

Removing these causes Errai to use its built-in
converters which do handle the empty string.
